requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
schedule==1.0.0
soupsieve==2.2.1
urllib3==1.26.4
python-dotenv==1.0.0
pandas==2.1.1

